Separate some base code
